% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCLUE.R
\name{ComputeSepScores}
\alias{ComputeSepScores}
\title{Compute a separation score between different clusters}
\usage{
ComputeSepScores(log_scdata = in_data, membership, topK)
}
\arguments{
\item{membership}{Membership index for the clustering results}

\item{topK}{Top K highly variable genes}

\item{data}{The gene expression matrix in log scale. Each row corresponds to the gene and column corresponds to the cell}
}
\description{
ComputeSepScores computes the separation score between different clsters using Welch's t-test
}
