% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCLUE.R
\name{make_dsm}
\alias{make_dsm}
\title{Make a doubly stochastic matrix}
\usage{
make_dsm(my_mat)
}
\arguments{
\item{my_may}{Square matrix}
}
\description{
make_dsm converts a square matrix into a doubly stochastic matrix.
}
