% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCLUE.R
\name{MakeKNN}
\alias{MakeKNN}
\title{Construct a KNN (K nearest neighbors) network}
\usage{
MakeKNN(data = in_data, k_num = in_k)
}
\arguments{
\item{data}{The gene expression matrix in log scale. Each row corresponds to the gene and column corresponds to the cell}

\item{k_num}{The number of nearest neighbors to construct KNN network}
}
\description{
Construct a KNN (K nearest neighbors) network
}
