% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCLUE.R
\name{MergeClusters}
\alias{MergeClusters}
\title{Merge different clusters based on a separation score}
\usage{
MergeClusters(membership, sep_scores)
}
\arguments{
\item{membership}{Membership index for the clustering results}

\item{sep_scores}{Separation scores between each clusters}
}
\description{
MergeClusters iteratively merges different clusters based on the separation scores.
}
