% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCLUE.R
\name{scCLUE}
\alias{scCLUE}
\title{scCLUE: single cell clustering through ensemble feature selection and similarity measurement}
\usage{
scCLUE(log_scdata = data, nCls, nEns = 15, K = c(5, 10), nPCs = 15,
  smin = 0.6, smax = 1)
}
\arguments{
\item{log_scdata}{(M x N) dimensional matrix for gene expressions (log scale)
M: The number of genes
N: The number of cells}

\item{nCls}{The number of clusters}

\item{nEns}{The number of trials for ensemble feature selection and similarity measurements}

\item{K}{The number of nearest neighbors for KNN networks}

\item{smin}{Ratio of the minimum feature sampling}

\item{smax}{Ratio of the maximum feature sampling}
}
\description{
scCLUE: single cell clustering through ensemble feature selection and similarity measurement
}
\references{
Hyundoo Jeong

Effective single cell clustering through ensemble feature selection and similarity measurements
}
\author{
Hyundoo Jeong
}
